Fix/signed int for mint (#57)
* fix: fix signed int for mint

* fix: fix example

* pass script mint amounts as string to csl

* comment maestro test for now

* fix: update ci script with env

* chore: revert test case

* test: testing ci

* test: testing ci

* test: testing ci

---------

+++ packages/sidan-csl-rs/src/model/tx_builder_types/mint.rs
@@ -29,5 +29,5 @@ pub struct SimpleScriptMint {
 pub struct MintParameter {
     pub policy_id: String,
     pub asset_name: String,
-    pub amount: u64,
+    pub amount: i128,
 }
